Update 2.7.tex
I kept getting errors in MikTex so I'm playing it safe with the brackets for now. It also seems to handle the soft-wrap nicely, cos I'm not having any problems with mal-assigned line-breaks - this also means that it's not obvious when LineBreaks are needed. How's it look now?

Thanks for spotting the iff mistake - was accidentally doing this proof for complete bipartite graphs.